refactor: 패키지 이름 수정
패키지 이름이 대문자로 되어있던 부분을 수정한다.
