+++ src/main/java/com/splanet/splanet/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.splanet.splanet.Comment.repository;
+package com.splanet.splanet.comment.repository;
 
-import com.splanet.splanet.Comment.entity.Comment;
+import com.splanet.splanet.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
