+++ src/main/java/com/splanet/splanet/friendRequest/entity/FriendRequest.java
@@ -1,11 +1,9 @@
-package com.splanet.splanet.Payment.entity;
+package com.splanet.splanet.friendRequest.entity;
 
 import com.splanet.splanet.core.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
-
-import java.time.LocalDateTime;
 
 @SuperBuilder
 @Getter
@@ -13,24 +11,21 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Payment extends BaseEntity {
+public class FriendRequest extends BaseEntity {
 
-    @Column(name = "subscription_id", nullable = false)
-    private Long subscriptionId;
+    @Column(name = "requester_id", nullable = false)
+    private Long requesterId;
+
+    @Column(name = "receiver_id", nullable = false)
+    private Long receiverId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status")
     private Status status;
 
-    @Column(name = "price")
-    private int price;
-
-    @Column(name = "payment_date")
-    private LocalDateTime paymentDate;
-
     public enum Status {
-        COMPLETED,
         PENDING,
-        FAILED
+        ACCEPTED,
+        REJECTED
     }
 }
